+++ dataworkspace/dataworkspace/tests/datasets/test_utils.py
@@ -1,0 +1,21 @@
+from dataworkspace.apps.datasets.models import DataSet
+from dataworkspace.apps.datasets.utils import (
+    dataset_type_to_manage_unpublished_permission_codename,
+)
+
+
+def test_dataset_type_to_manage_unpublished_permission_codename():
+    assert (
+        dataset_type_to_manage_unpublished_permission_codename(0)
+        == 'datasets.manage_unpublished_reference_datasets'
+    )
+    assert (
+        dataset_type_to_manage_unpublished_permission_codename(DataSet.TYPE_DATA_CUT)
+        == 'datasets.manage_unpublished_datacut_datasets'
+    )
+    assert (
+        dataset_type_to_manage_unpublished_permission_codename(
+            DataSet.TYPE_MASTER_DATASET
+        )
+        == 'datasets.manage_unpublished_master_datasets'
+    )